force int to price

+++ page/cdiscount.py
@@ -75,7 +75,7 @@ class Cdiscount(Crawl):
 						'product_id' : int(product_id),
 						'name'  : product_name,
 						'image' : product_image,
-						'price' : price,
+						'price' : (int)price,
 						'url'   : url
 					}
